auth/requirements.txt
Flask==1.1.1
Flask-API.yandex==0.6.2.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
flask-login==0.5.0
Flask-script
urllib3==1.25.8
requests==2.23.0
mysqlclient
flask_expects_json==1.4.0
pika==1.1.0
PyJWT==1.7.1
